Class names follows command names now
